move client to dedicated file
